requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# MQTT
paho-mqtt>=1.6.1
# Alternative: asyncio-mqtt>=0.16.1

# HTTP client (for GitHub API, etc.)
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0

# Numerical computation (required for simulator)
numpy>=1.26.0

# Optional: For advanced anomaly detection
# scikit-learn>=1.3.0
